增加怪物技能配置，完善怪物攻击和技能，优化部分资源 下一步：玩家死亡处理

+++ GDServer/GDClient.cs
@@ -35,7 +35,7 @@ namespace GDServer {
         internal void BeAttacked(int damage) {
             if (m_IsDie)
                 return;
-            FP.FightHP -= (short)damage;
+            FP.FightHP -= damage;
             if (FP.FightHP <=0&& !m_IsDie) {
                 m_IsDie = true;
             }
